+++ srcs/lin_alg/vec3_operations2.c
@@ -6,7 +6,7 @@
 /*   By: dshatilo <dshatilo@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/27 10:17:57 by dshatilo          #+#    #+#             */
-/*   Updated: 2024/07/24 22:56:25 by dshatilo         ###   ########.fr       */
+/*   Updated: 2024/07/29 14:51:10 by dshatilo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 float	vec3_norm(t_vec3 vect)
 {
-	return (sqrt(vec3_dot(vect, vect)));
+	return (sqrtf(vec3_dot(vect, vect)));
 }
 
 t_vec3	vec3_unit(t_vec3 vect)
@@ -41,11 +41,11 @@ t_vec3	vec3_cross(t_vec3 a, t_vec3 b)
 
 bool	vec3_equals(t_vec3 *a, t_vec3 *b)
 {
-	if (fabs(a->x - b->x) > 0.005)
+	if (fabsf(a->x - b->x) > 0.005)
 		return (false);
-	if (fabs(a->y - b->y) > 0.005)
+	if (fabsf(a->y - b->y) > 0.005)
 		return (false);
-	if (fabs(a->z - b->z) > 0.005)
+	if (fabsf(a->z - b->z) > 0.005)
 		return (false);
 	return (true);
 }
@@ -54,6 +54,6 @@ float	vec3_distance(t_vec3 a, t_vec3 b)
 {
 	float	d;
 
-	d = sqrt(pow(b.x - a.x, 2) + pow(b.y - a.y, 2) + pow(b.z - a.z, 2));
+	d = sqrtf(powf(b.x - a.x, 2) + powf(b.y - a.y, 2) + powf(b.z - a.z, 2));
 	return (d);
 }
